# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Itinera Routing API
  version: 0.1.0
  description: Turn-by-turn routing engine with contraction hierarchies.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    email: tiletopia@proton.me
servers:
  - url: http://localhost:5000
    description: Default bind of `itinera serve`
  - url: http://localhost:3000
    description: Default bind of the Docker image
security:
  - bearerAuth: []
paths:
  /route:
    get:
      summary: Calculate route
      description: Find the fastest path between two coordinates.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: "Source coordinate as lat,lon"
          example: "51.5074,-0.1278"
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: "Target coordinate as lat,lon"
          example: "51.5085,-0.0762"
        - name: profile
          in: query
          schema:
            $ref: '#/components/schemas/Profile'
        - name: algorithm
          in: query
          schema:
            type: string
            enum: [dijkstra, astar, ch]
            default: astar
          description: "`ch` requires a preprocessed contraction hierarchy"
      responses:
        '200':
          description: Route result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No route between the two nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /nearest:
    get:
      summary: Snap to nearest node
      description: Find the nearest node on the road network.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lon
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Nearest node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /isochrone:
    get:
      summary: Generate isochrone
      description: Calculate the area reachable from a point within a time budget.
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lon
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: max_seconds
          in: query
          required: true
          schema:
            type: number
            format: double
          description: Travel time budget in seconds
        - name: profile
          in: query
          schema:
            $ref: '#/components/schemas/Profile'
      responses:
        '200':
          description: Reachable area
          content:
            application/json:
              schema:
                type: object
                properties:
                  reachable_nodes:
                    type: integer
                  boundary:
                    type: array
                    description: Convex hull of the reachable nodes
                    items:
                      $ref: '#/components/schemas/LatLonPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /match:
    post:
      summary: Map match a GPS trace
      description: >-
        Snap a noisy GPS trace onto the loaded routing graph with a hidden Markov
        model. The graph is indexed as one segment per road, so a two-way road
        counts once. Coordinates in the response are [lon, lat]. Only `lat` and
        `lon` steer the match, the other trace point fields are accepted and unused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trace]
              properties:
                trace:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items:
                    $ref: '#/components/schemas/TracePoint'
                profile:
                  $ref: '#/components/schemas/MatchProfile'
                search_radius_m:
                  type: number
                  format: double
                  default: 50
                  exclusiveMinimum: 0
                  maximum: 1000
                  description: How far from a trace point to look for roads
      responses:
        '200':
          description: The matched trace
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  matched_points:
                    type: array
                    description: One per trace point, in request order
                    items:
                      type: object
                      properties:
                        original:
                          $ref: '#/components/schemas/LonLatPair'
                        snapped:
                          $ref: '#/components/schemas/LonLatPair'
                        distance_from_road_m:
                          type: number
                        road_name:
                          type: [string, 'null']
                          description: Null where no road was within the search radius
                  matched_route:
                    type: array
                    items:
                      $ref: '#/components/schemas/LonLatPair'
                  confidence:
                    type: number
                    description: 0 to 1, the mean fit of the matched points
                  total_distance_m:
                    type: number
                  road_segments:
                    type: array
                    description: Roads the trace ran along, longest first
                    items:
                      type: object
                      properties:
                        road_name:
                          type: string
                        road_class:
                          type: string
                        distance_m:
                          type: number
                        duration_secs:
                          type: number
                        speed_kmh:
                          type: number
        '400':
          description: Empty or oversized trace (max 1000 points), or a bad profile or radius
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /network/components:
    post:
      summary: Connected components
      description: Weakly connected components of the loaded graph.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                top_k:
                  type: integer
                  default: 20
                  minimum: 1
                  maximum: 1000
                  description: Number of components to return, largest first
      responses:
        '200':
          description: Component summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  num_nodes:
                    type: integer
                  num_components:
                    type: integer
                  largest_component_size:
                    type: integer
                  components:
                    type: array
                    items:
                      type: object
                      properties:
                        component_id:
                          type: integer
                        size:
                          type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Enforced only when the server is started with ITINERA_JWT_SECRET set,
        otherwise every route is open.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token. The body is empty.
  schemas:
    Profile:
      type: string
      enum: [car, bicycle, pedestrian, truck]
      default: car
...
    LatLonPair:
      type: array
      description: "[lat, lon]"
      minItems: 2
      maxItems: 2
      items:
        type: number
        format: double
    LonLatPair:
      type: array
      description: "[lon, lat]"
      minItems: 2
      maxItems: 2
      items:
        type: number
        format: double
    MatchProfile:
      type: string
      enum: [driving, walking, cycling]
      default: driving
    TracePoint:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        timestamp:
          type: number
          format: double
          description: Seconds, any epoch
        accuracy_m:
          type: number
          format: double
        speed_mps:
          type: number
          format: double
        bearing_deg:
          type: number
          format: double
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    RouteResponse:
      type: object
      properties:
        distance_m:
          type: number
          description: Total distance in meters
        duration_s:
          type: number
          description: Total travel time in seconds
        geometry:
          type: array
          items:
            $ref: '#/components/schemas/LatLonPair'
        steps:
          type: array
          description: Empty for the `ch` algorithm
          items:
            $ref: '#/components/schemas/RouteStep'
    RouteStep:
      type: object
      properties:
        distance_m:
          type: number
        duration_s:
          type: number
        name:
          type: [string, 'null']
          description: Street name, if the source data has one
        maneuver:
          type: string
          description: "Maneuver name, e.g. Depart, TurnRight, Arrive"
    NearestResponse:
      type: object
      properties:
        node_id:
          type: integer
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        distance_m:
          type: number
          description: Distance from the query point in meters